requirements.txt
absl-py==2.1.0
alembic==1.14.0
attrs==24.3.0
blinker==1.9.0
cffi==1.17.1
click==8.1.8
contourpy==1.3.1
cycler==0.12.1
dlib==19.24.6
face-recognition==1.3.0
face_recognition_models==0.3.0
Flask==3.1.0
Flask-Migrate==4.1.0
Flask-SQLAlchemy==3.1.1
flatbuffers==24.12.23
fonttools==4.55.3
greenlet==3.1.1
itsdangerous==2.2.0
jax==0.5.0
jaxlib==0.5.0
Jinja2==3.1.5
kiwisolver==1.4.8
Mako==1.3.8
MarkupSafe==3.0.2
matplotlib==3.10.0
mediapipe==0.10.20
ml_dtypes==0.5.1
numpy==1.26.4
opencv-contrib-python==4.11.0.86
opt_einsum==3.4.0
packaging==24.2
pillow==11.1.0
protobuf==4.25.5
pycparser==2.22
pyparsing==3.2.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
scipy==1.15.1
sentencepiece==0.2.0
six==1.17.0
sounddevice==0.5.1
SQLAlchemy==2.0.37
typing_extensions==4.12.2
Werkzeug==3.1.3
